+++ tools/squids2vtk.cpp
@@ -24,12 +24,12 @@ int main( int argc, char** argv)
     fprintf(f,"vtk output\n");
     fprintf(f,"ASCII\n");
     fprintf(f,"DATASET POLYDATA\n");
-    fprintf(f,"POINTS %d float\n",squids.nlin());
+    fprintf(f,"POINTS %d float\n",(int)squids.nlin());
     for( unsigned int i = 0; i < squids.nlin(); i += 1 )
     {
         fprintf(f, "%f %f %f\n", squids(i,0), squids(i,1), squids(i,2));
     }
-    fprintf(f,"POINT_DATA %d\n",squids.nlin());
+    fprintf(f,"POINT_DATA %d\n",(int)squids.nlin());
     fprintf(f,"NORMALS normals float\n");
     for( unsigned int i = 0; i < squids.nlin(); i += 1 )
     {
